requirements.txt
Flask==2.3.2
twilio==8.3.0
openai==0.27.4
requests==2.31.0
python-dotenv==1.0.0
